# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3341",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api-docs": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/contacts/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get All Contacts",
            "schema": {
              "$ref": "#/definitions/Contacts"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "contact",
            "in": "body",
            "description": "Create a new contact",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/contacts/{contactId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get One Contact by ID",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "contact",
            "in": "body",
            "description": "Update a single contact",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Updated successfully"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Josh"
        },
        "lastName": {
          "type": "string",
          "example": "Allen"
        },
        "email": {
          "type": "string",
          "example": "joshallen@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "blue"
        },
        "birthday": {
          "type": "string",
          "example": "1995-05-22T07:00:00.000Z"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email",
        "favoriteColor",
        "birthday"
      ]
    },
    "Contacts": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Josh"
          },
          "lastName": {
            "type": "string",
            "example": "Allen"
          },
          "email": {
            "type": "string",
            "example": "joshallen@gmail.com"
          },
          "favoriteColor": {
            "type": "string",
            "example": "blue"
          },
          "birthday": {
            "type": "string",
            "example": "1995-05-22T07:00:00.000Z"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "favoriteColor",
          "birthday"
        ]
      }
    }
  }
}